redshift_lambda/requirements.txt
aws-psycopg2==1.2.1
python-dotenv==0.14.0
